+++ src/drawer/latex.rs
@@ -32,7 +32,7 @@ use crate::projector::{
 };
 use crate::script::HashableArc;
 
-use crate::script::figure::Style::{self, Bolded, Dashed, Solid, Dotted};
+use crate::script::figure::Style::{self, Bold, Dashed, Dotted, Solid};
 
 /// Function getting the point's name (if it exists, if not then it returns the point's coordinates).
 fn get_point_name(
@@ -58,7 +58,7 @@ fn assign_mode(rendered: &Rendered, mode: Style) -> String {
         _ => match mode {
             Dotted => "dotted".to_string(),
             Dashed => "dashed".to_string(),
-            Bolded => "ultra thick".to_string(),
+            Bold => "ultra thick".to_string(),
             Solid => "thin".to_string(),
         },
     }
@@ -73,7 +73,13 @@ fn points(point: &Rc<RenderedPoint>, scale: f64) -> String {
             \coordinate ({}) at ({}, {}); \fill[black] ({}) circle (1pt);
             \node at ({}, {}) {{{}}};
         "#,
-        point.uuid, position.real, position.imaginary, point.uuid, label_position.real, label_position.imaginary, point.label
+        point.uuid,
+        position.real,
+        position.imaginary,
+        point.uuid,
+        label_position.real,
+        label_position.imaginary,
+        point.label
     )
 }