+++ src/labels.rs
@@ -20,7 +20,13 @@ CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 use std::f64::consts::PI;
 
-use crate::{projector::{RenderedLine, RenderedAngle, RenderedRay, RenderedSegment, RenderedCircle}, generator::{geometry::{distance_pt_ln, get_line, distance_pt_pt, rotate_around}, Complex}};
+use crate::{
+    generator::{
+        geometry::{distance_pt_ln, distance_pt_pt, get_line, rotate_around},
+        Complex,
+    },
+    projector::{RenderedAngle, RenderedCircle, RenderedLine, RenderedRay, RenderedSegment},
+};
 
 /// Function that outputs the position of the label. The position is located on the bisector of the given angle.
 /// ///
@@ -28,15 +34,18 @@ use crate::{projector::{RenderedLine, RenderedAngle, RenderedRay, RenderedSegmen
 /// Panics when the angle is equal to 0, which shouldn't happen.  
 fn label_position(vec: Complex, vec_next: Complex, angle: f64, point: Complex) -> Complex {
     // We get the the bisector angle.
-    let bisector_angle = ((vec_next.arg() - vec.arg()).rem_euclid(2.0 * PI))/2.0 + vec.arg();
+    let bisector_angle = ((vec_next.arg() - vec.arg()).rem_euclid(2.0 * PI)) / 2.0 + vec.arg();
 
     // This is just the standard complex number formula.
     let bisector_vec = Complex::new(bisector_angle.cos(), bisector_angle.sin()).normalize();
 
     // to do -> better scaling
-    let scale = 540.0/angle.to_degrees().abs();
+    let scale = 540.0 / angle.to_degrees().abs();
 
-    Complex::new(point.real + (bisector_vec.real * 3.0 * scale), point.imaginary + (bisector_vec.imaginary * 3.0 * scale))
+    Complex::new(
+        point.real + (bisector_vec.real * 3.0 * scale),
+        point.imaginary + (bisector_vec.imaginary * 3.0 * scale),
+    )
 }
 
 /// Function that gets the label position.
@@ -81,16 +90,16 @@ fn get_label_pos(vec_associated: &Vec<Complex>, point: Complex) -> Complex {
 /// Function that output the position of the label.
 /// ///
 /// # Panics
-/// 
+///
 #[must_use]
 pub fn point_label_position(
-    lines: &Vec<RenderedLine>, 
-    angles: &Vec<RenderedAngle>, 
+    lines: &Vec<RenderedLine>,
+    angles: &Vec<RenderedAngle>,
     segments: &Vec<RenderedSegment>,
     rays: &Vec<RenderedRay>,
     circles: &Vec<RenderedCircle>,
     mut vec_associated: Vec<Complex>,
-    point: Complex
+    point: Complex,
 ) -> Complex {
     // Checking the lines for associated vectors.
     for ln in lines {
@@ -133,7 +142,11 @@ pub fn point_label_position(
 
         if distance < 1e-2 {
             if a > 0.0 {
-                if seg1_mov.real - u1 < point.real && point.real < seg2_mov.real + u1 && seg1_mov.imaginary - u2 < point.imaginary && point.imaginary < seg2_mov.imaginary + u2 {
+                if seg1_mov.real - u1 < point.real
+                    && point.real < seg2_mov.real + u1
+                    && seg1_mov.imaginary - u2 < point.imaginary
+                    && point.imaginary < seg2_mov.imaginary + u2
+                {
                     if distance_pt_pt(point, seg1) < 1.0 {
                         vec_associated.push(ln.direction);
                     } else if distance_pt_pt(point, seg2) < 1.0 {
@@ -143,7 +156,12 @@ pub fn point_label_position(
                         vec_associated.push(-ln.direction);
                     }
                 }
-            } else if a < 0.0 && seg1_mov.real - u1 < point.real && point.real < seg2_mov.real + u1 && seg1_mov.imaginary - u2 > point.imaginary && point.imaginary > seg2_mov.imaginary + u2 {
+            } else if a < 0.0
+                && seg1_mov.real - u1 < point.real
+                && point.real < seg2_mov.real + u1
+                && seg1_mov.imaginary - u2 > point.imaginary
+                && point.imaginary > seg2_mov.imaginary + u2
+            {
                 if distance_pt_pt(point, seg1) < 1.0 {
                     vec_associated.push(ln.direction);
                 } else if distance_pt_pt(point, seg2) < 1.0 {
@@ -159,11 +177,13 @@ pub fn point_label_position(
     for ray in rays {
         let frame_point = ray.points.1;
 
-        let ray_point = if distance_pt_pt(ray.points.0, frame_point) > distance_pt_pt(ray.draw_point, frame_point) {
+        let ray_point = if distance_pt_pt(ray.points.0, frame_point)
+            > distance_pt_pt(ray.draw_point, frame_point)
+        {
             ray.points.0
         } else {
             ray.draw_point
-        }; 
+        };
 
         let ln = get_line(ray_point, frame_point);
         let a = ln.direction.real * ln.direction.imaginary;
@@ -172,12 +192,28 @@ pub fn point_label_position(
         let u1 = unit.real;
         let u2 = unit.imaginary;
 
-        let b1 = ray_point.real > frame_point.real && frame_point.real < point.real && point.real < ray_point.real - u1 && frame_point.imaginary < point.imaginary && point.imaginary < ray_point.imaginary - u2;
-        let b2 = frame_point.real > ray_point.real && ray_point.real - u1 < point.real && point.real < frame_point.real && ray_point.imaginary - u2 < point.imaginary && point.imaginary < frame_point.imaginary;
+        let b1 = ray_point.real > frame_point.real
+            && frame_point.real < point.real
+            && point.real < ray_point.real - u1
+            && frame_point.imaginary < point.imaginary
+            && point.imaginary < ray_point.imaginary - u2;
+        let b2 = frame_point.real > ray_point.real
+            && ray_point.real - u1 < point.real
+            && point.real < frame_point.real
+            && ray_point.imaginary - u2 < point.imaginary
+            && point.imaginary < frame_point.imaginary;
 
-        let b3 = frame_point.real > ray_point.real && ray_point.real - u1 < point.real && point.real < frame_point.real && point.imaginary < ray_point.imaginary - u2 && frame_point.imaginary < point.imaginary;
-        let b4 = ray_point.real > frame_point.real && frame_point.real < point.real && point.real < ray_point.real - u1 && point.imaginary < frame_point.imaginary && ray_point.imaginary - u2 < point.imaginary;
-            
+        let b3 = frame_point.real > ray_point.real
+            && ray_point.real - u1 < point.real
+            && point.real < frame_point.real
+            && point.imaginary < ray_point.imaginary - u2
+            && frame_point.imaginary < point.imaginary;
+        let b4 = ray_point.real > frame_point.real
+            && frame_point.real < point.real
+            && point.real < ray_point.real - u1
+            && point.imaginary < frame_point.imaginary
+            && ray_point.imaginary - u2 < point.imaginary;
+
         if distance_pt_ln(point, ln) < 1e-2 {
             if a > 0.0 {
                 if b1 || b2 {
@@ -208,7 +244,7 @@ pub fn point_label_position(
         }
     }
 
-    vec_associated.sort_by(|a,b| (a.arg()).partial_cmp(&b.arg()).unwrap());
+    vec_associated.sort_by(|a, b| (a.arg()).partial_cmp(&b.arg()).unwrap());
 
     get_label_pos(&vec_associated, point)
 }